+++ speedd-ml/src/main/scala/org/speedd/ml/model/fz/package.scala
@@ -28,6 +28,10 @@
 
 package org.speedd.ml.model
 
-object FZ {
-// todo
+package object fz {
+
+  /**
+   * The Cassandra keyspace to use
+   */
+  val KEYSPACE = "fz"
 }
